requirements.txt
opencv-python
numpy
scikit-learn
pillow
tqdm
insightface
onnxruntime
matplotlib
pandas